Splitter opp scenarioer og lager readme
