+++ formidling/src/test/java/no/nav/ung/sak/formidling/ManuellVedtaksbrevTest.java
@@ -10,7 +10,7 @@ import no.nav.ung.sak.behandlingslager.behandling.aksjonspunkt.AksjonspunktTestS
 import no.nav.ung.sak.behandlingslager.behandling.repository.BehandlingRepository;
 import no.nav.ung.sak.behandlingslager.formidling.VedtaksbrevValgEntitet;
 import no.nav.ung.sak.behandlingslager.formidling.VedtaksbrevValgRepository;
-import no.nav.ung.sak.formidling.scenarioer.BrevScenarioer;
+import no.nav.ung.sak.formidling.scenarioer.EndringInntektScenarioer;
 import no.nav.ung.sak.formidling.vedtak.regler.VedtaksbrevInnholdbyggerStrategy;
 import no.nav.ung.sak.test.util.behandling.TestScenarioBuilder;
 import no.nav.ung.sak.test.util.behandling.UngTestScenario;
@@ -47,7 +47,7 @@ class ManuellVedtaksbrevTest extends AbstractVedtaksbrevInnholdByggerTest {
         LocalDate fom = LocalDate.of(2024, 12, 1);
 
         var behandling = lagScenario(
-            BrevScenarioer.endring0KrInntekt_19år(fom), AksjonspunktDefinisjon.FORESLÅ_VEDTAK_MANUELT);
+            EndringInntektScenarioer.endring0KrInntekt_19år(fom), AksjonspunktDefinisjon.FORESLÅ_VEDTAK_MANUELT);
 
         vedtaksbrevValgRepository.lagre(new VedtaksbrevValgEntitet(
             behandling.getId(),
@@ -108,7 +108,7 @@ class ManuellVedtaksbrevTest extends AbstractVedtaksbrevInnholdByggerTest {
     @Override
     protected Behandling lagScenarioForFellesTester() {
         LocalDate fom = LocalDate.of(2024, 12, 1);
-        Behandling behandling = lagScenario(BrevScenarioer.endring0KrInntekt_19år(fom), AksjonspunktDefinisjon.FORESLÅ_VEDTAK_MANUELT);
+        Behandling behandling = lagScenario(EndringInntektScenarioer.endring0KrInntekt_19år(fom), AksjonspunktDefinisjon.FORESLÅ_VEDTAK_MANUELT);
 
         vedtaksbrevValgRepository.lagre(new VedtaksbrevValgEntitet(
             behandling.getId(),
