+++ formidling/src/test/java/no/nav/ung/sak/formidling/informasjonsbrev/InformasjonsbrevVerifikasjon.java
@@ -1,7 +1,7 @@
 package no.nav.ung.sak.formidling.informasjonsbrev;
 
 import no.nav.ung.sak.formidling.BrevTestUtils;
-import no.nav.ung.sak.formidling.scenarioer.BrevScenarioer;
+import no.nav.ung.sak.formidling.scenarioer.BrevScenarioerUtils;
 
 import java.time.LocalDate;
 
@@ -22,7 +22,7 @@ public class InformasjonsbrevVerifikasjon {
         return STANDARD_HEADER_FOOTER
             .formatted(
                 BrevTestUtils.brevDatoString(brevdato),
-                BrevScenarioer.DEFAULT_NAVN,
+                BrevScenarioerUtils.DEFAULT_NAVN,
                 fnr,
                 body);
     }
